# servidor_fastapi/openapi.yaml
openapi: 3.0.3
info:
  description: "API para gestionar usuarios (no logeados, logueados, artista), productos\
    \ y compras."
  title: Music E-Commerce API
  version: 1.0.0
servers:
- description: Servidor local de FastAPI
  url: http://localhost:8000
security:
- BearerAuth: []
...
paths:
  /auth/register:
    post:
      description: Crea un nuevo usuario en la base de datos (Firebase + la BD local).
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: Usuario registrado correctamente.
        "400":
          description: Error de validación o usuario ya existente.
      summary: Registrar un nuevo usuario (sea normal o artista)
      tags:
      - Auth
  /auth/login:
    post:
      description: Devuelve un token JWT si las credenciales son correctas.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_auth_login_post_200_response'
          description: Inicio de sesión exitoso
        "401":
          description: Credenciales inválidas
      summary: Iniciar sesión
      tags:
      - Auth
...
components:
  schemas:
    UserRegisterInput:
      example:
        password: password
        role: role
        displayName: displayName
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          title: password
          type: string
        displayName:
          title: displayName
          type: string
        role:
          description: "Ejemplo: 'user' o 'artista'"
          title: role
          type: string
      required:
      - email
      - password
      title: UserRegisterInput
      type: object
    UserLoginInput:
      example:
        password: password
        email: email
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          title: password
          type: string
      required:
      - email
      - password
      title: UserLoginInput
      type: object
    UserProfile:
      example:
        role: role
        displayName: displayName
        imageUrl: https://openapi-generator.tech
        id: id
        email: email
      properties:
        id:
          description: Identificador único del usuario
          title: id
          type: string
        email:
          format: email
          title: email
          type: string
        displayName:
          title: displayName
          type: string
        role:
          description: user | artista
          title: role
          type: string
        imageUrl:
          description: Link a la imagen del perfil del usuario/artista
          format: uri
          title: imageUrl
          type: string
      title: UserProfile
      type: object
...
    _auth_login_post_200_response:
      example:
        token: token
      properties:
        token:
          description: Token JWT para autenticar futuras peticiones
          title: token
          type: string
      title: _auth_login_post_200_response
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http